Refactor: Standardize component widths and theme usage
This commit refactors several UI components to use `fillMaxWidth` instead of fixed widths, ensuring they adapt to different screen sizes. It also updates components to consistently use `AppTheme` for styling, specifically for colors, text styles, and spacing.

Key changes:

- **Width Adjustment:**
    - `SearchScreen`: `SearchTextInputField`, `RecentSearchHeader`, and `RecentSearchList` now use `fillMaxWidth()` instead of a fixed `width(360.dp)`.
- **Theme Consistency:**
    - `RecentSearchHeader`: Now uses `AppTheme.radius.medium` for corner rounding.
    - `SearchTextInputField`: Now uses `AppTheme.spacing.medium` for horizontal padding and `AppTheme.radius.medium` for corner rounding.
- **Content Description:**
    - `RecentSearchItem`: Updated content descriptions for "History Icon" and "Remove Icon" to be more specific.
- **Import Adjustments:**
    - `RecentSearchList.kt`: Package path updated from `com.example.presentation.component.search` to `com.example.presentation.component`.
    - `RecentSearchScreen.kt`: The `items` parameter for `RecentSearchList` now correctly maps `RecentSearchItem` to `item.query`.
    - Minor import changes across files to reflect the movement of `RecentSearchList` and to remove unused imports.

The `SearchListViewModel.kt` files in both `presentation/src/main/java/com/example/presentation/component/screens/searchscreen/` and `presentation/src/main/java/com/example/presentation/component/searchListUi/` were updated to import `RecentSearchItem` from `com.example.domain` instead of `com.example.presentation.component.search`.
