+++ presentation/src/main/java/com/example/presentation/component/screens/searchscreen/SearchListViewModel.kt
@@ -1,45 +1,33 @@
 package com.example.presentation.component.screens.SearchScreen
 
-
+import com.example.data.InMemoryRecentSearchRepository
+import com.example.presentation.component.search.RecentSearchItem
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import androidx.lifecycle.ViewModel
-import com.example.domain.RecentSearchRepository
 import com.example.domain.RecentSearchItem
-import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.launch
 
-class SearchListViewModel(
-    private val repo: RecentSearchRepository
-) : ViewModel() {
+class SearchListViewModel(private val repo: InMemoryRecentSearchRepository) : ViewModel() {
     private val _recentSearches = MutableStateFlow<List<RecentSearchItem>>(emptyList())
     val recentSearches: StateFlow<List<RecentSearchItem>> = _recentSearches.asStateFlow()
 
     fun loadRecentSearches() {
-        viewModelScope.launch {
-            _recentSearches.value = repo.getRecentSearches()
-        }
+        _recentSearches.value = repo.getRecentSearches()
     }
 
     fun addRecentSearch(item: RecentSearchItem) {
-        viewModelScope.launch {
-            repo.addRecentSearch(item)
-            loadRecentSearches()
-        }
+        repo.addRecentSearch(item)
+        _recentSearches.value = repo.getRecentSearches()
     }
 
-    fun removeRecentSearch(item: RecentSearchItem) {
-        viewModelScope.launch {
-            repo.removeRecentSearch(item.id.toString())
-            loadRecentSearches()
-        }
+    fun removeRecentSearch(id: String) {
+        repo.removeRecentSearch(id)
+        _recentSearches.value = repo.getRecentSearches()
     }
 
-    fun clearAllRecentSearches() {
-        viewModelScope.launch {
-            repo.clearAllRecentSearches()
-            loadRecentSearches()
-        }
+    fun clearAll() {
+        repo.clearAllRecentSearches()
+        _recentSearches.value = repo.getRecentSearches()
     }
 }
